refactor: project structure and ts paths
