+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Navbar from "@components/Navbar";
+import Navbar from "@/components/navbar";
 
 export default function Home() {
   return (
