+++ tsconfig.json
@@ -20,10 +20,7 @@
     ],
     "baseUrl": "./src/",
     "paths": {
-      "@components/*": ["app/components/*"],
-      "@styles/*": ["app/styles/*"],
-      "@api/*": ["app/api/*"],
-      "@database/*": ["database/*"]
+      "@/*": ["*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
